# src/stacks/api/types/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: `${string}://${string}` | (string & {});
};

export type LocationRecord = {
    id: string;
    name: string;
    latitude: number;
    longitude: number;
    country: string;
    /**
     * State or county of the city
     */
    region: string;
    city: string;
};

export type LocationData = {
    data: Array<LocationRecord>;
};

export type Error = {
    /**
     * Stable, machine-readable error identifier.
     */
    code: string;
    message: string;
};

export type ErrorNoCode = {
    message: string;
};

/**
 * Unique location identifier provided on record creation
 */
export type LocationId = string;

export type ListLocationRecordsData = {
    body?: never;
    path?: never;
    query?: {
        /**
         * Maximum number of location records to return
         */
        limit?: number;
    };
    url: '/locations';
};

export type ListLocationRecordsErrors = {
    /**
     * Invalid request format
     */
    400: Error;
    /**
     * Unauthorized request
     */
    403: ErrorNoCode;
    /**
     * An unexpected error occurred on the server
     */
    500: Error;
};

export type ListLocationRecordsError = ListLocationRecordsErrors[keyof ListLocationRecordsErrors];

export type ListLocationRecordsResponses = {
    /**
     * Successfully retrieved locations
     */
    200: LocationData;
};

export type ListLocationRecordsResponse = ListLocationRecordsResponses[keyof ListLocationRecordsResponses];

export type CreateLocationRecordData = {
    body: {
        id: string;
        name: string;
        country: string;
        /**
         * The administrative region in which the city is located.
         * Due to limitations in OpenStreetMap's administrative hierarchy,
         * this must be the administrative region in which the city is located.
         *
         */
        region: string;
        city: string;
    };
    path?: never;
    query?: never;
    url: '/locations';
};

export type CreateLocationRecordErrors = {
    /**
     * Invalid request format
     */
    400: Error;
    /**
     * Unauthorized request
     */
    403: ErrorNoCode;
    /**
     * A location with this id already exists
     */
    409: Error;
    /**
     * An unexpected error occurred on the server
     */
    500: Error;
};

export type CreateLocationRecordError = CreateLocationRecordErrors[keyof CreateLocationRecordErrors];

export type CreateLocationRecordResponses = {
    /**
     * Successfully created location
     */
    201: LocationData;
};

export type CreateLocationRecordResponse = CreateLocationRecordResponses[keyof CreateLocationRecordResponses];

export type DeleteLocationData = {
    body?: never;
    path: {
        /**
         * Unique location identifier provided on record creation
         */
        locationId: string;
    };
    query?: never;
    url: '/locations/{locationId}';
};

export type DeleteLocationErrors = {
    /**
     * Unauthorized request
     */
    403: ErrorNoCode;
    /**
     * The requested resource was not found
     */
    404: Error;
    /**
     * An unexpected error occurred on the server
     */
    500: Error;
};

export type DeleteLocationError = DeleteLocationErrors[keyof DeleteLocationErrors];

export type DeleteLocationResponses = {
    /**
     * location has been deleted
     */
    204: void;
};

export type DeleteLocationResponse = DeleteLocationResponses[keyof DeleteLocationResponses];

export type GetLocationData = {
    body?: never;
    path: {
        /**
         * Unique location identifier provided on record creation
         */
        locationId: string;
    };
    query?: never;
    url: '/locations/{locationId}';
};

export type GetLocationErrors = {
    /**
     * Unauthorized request
     */
    403: ErrorNoCode;
    /**
     * The requested resource was not found
     */
    404: Error;
    /**
     * An unexpected error occurred on the server
     */
    500: Error;
};

export type GetLocationError = GetLocationErrors[keyof GetLocationErrors];

export type GetLocationResponses = {
    /**
     * Successfully retrieved location for target id
     */
    200: LocationData;
};

export type GetLocationResponse = GetLocationResponses[keyof GetLocationResponses];

export type UpdateLocationRecordData = {
    body: {
        name?: string;
        country?: string;
        region?: string;
        city?: string;
    };
    path: {
        /**
         * Unique location identifier provided on record creation
         */
        locationId: string;
    };
    query?: never;
    url: '/locations/{locationId}';
};

export type UpdateLocationRecordErrors = {
    /**
     * Invalid request format
     */
    400: Error;
    /**
     * Unauthorized request
     */
    403: ErrorNoCode;
    /**
     * The requested resource was not found
     */
    404: Error;
    /**
     * An unexpected error occurred on the server
     */
    500: Error;
};

export type UpdateLocationRecordError = UpdateLocationRecordErrors[keyof UpdateLocationRecordErrors];

export type UpdateLocationRecordResponses = {
    /**
     * Successfully updated location
     */
    200: LocationData;
};

export type UpdateLocationRecordResponse = UpdateLocationRecordResponses[keyof UpdateLocationRecordResponses];